♻ :: home hook 분리
