+++ src/Components/Home/NotSubscribed/RegisterSchool/RegisterSchool.tsx
@@ -1,19 +1,11 @@
 import * as S from '@/Components/Home/NotSubscribed/RegisterSchool/RegisterSchool.style';
-
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
 
+import useUnhome from '@/Hooks/HomeHook/UnHome/index';
 
 const RegisterSchool = () => {
-    const navigate = useNavigate();
-    const handleCreate = () => {
-        navigate('/createschool');
-    }
-
-    const handleJoin = () => {
-        navigate('/schoolcode ');
-    }
-
+    const { ...unHome } = useUnhome();
+    
     return (
         <S.RegisterSchoolContainer>
             <S.RegisterSchoolBox>
@@ -22,10 +14,10 @@ const RegisterSchool = () => {
                     <S.Subtitle> 학교를 등록한 뒤 스기를 사용할 수 있어요 </S.Subtitle>
                 </S.TextBox>
                 <S.ButtonBox>
-                    <S.NewSchoolButton onClick={handleCreate}>
+                    <S.NewSchoolButton onClick={unHome.handleCreate}>
                         <S.NewSchoolButtonText> 새 학교 만들기 </S.NewSchoolButtonText>
                     </S.NewSchoolButton>
-                    <S.JoinSchoolButton onClick={handleJoin}>
+                    <S.JoinSchoolButton onClick={unHome.handleJoin}>
                         <S.JoinSchoolButtonText> 기존 학교 가입 </S.JoinSchoolButtonText>
                     </S.JoinSchoolButton>
                 </S.ButtonBox>
